Major Adjustment to API_KEY File
Switching course to have the program build and store the URLs.

Added the base URL

Added default amount of questions

Added question type

This makes it so that if we want to adjust different settings it is much easier than having to go get a new link every time.

Added notes for the progress that will need to be completed in order for it to be functional.