Created Function to Fetch Questions from website
Within this function, it will take the URL that we received and will get the JSON text from the website, will clean up the text, and give it to us. We then shuffled the answers, and will have the cleaned up questions, answers, and location of the correct answer all within a list (prepared).